Move code that creates gTa from reptilian.scala to tokenization.scala
FIXME: Apparently cannot rename function (provisionally t(), but should be something like createGTa())
